model/requirements.txt
tensorboard-pytorch
tensorboard
pygraphviz
torchview
